fixed inner loop

+++ src/workflow.cpp
@@ -5,7 +5,7 @@
 Workflow::Workflow(
         std::vector<std::vector<int> > edges, 
         std::unordered_map<std::string, double> processingTimePerType,
-        std::unordered_map<int, int> dueDates,
+        std::unordered_map<int, double> dueDates,
         std::unordered_map<int, std::string> nodeNames
         ): dueDates(dueDates), nodeNames(nodeNames){
